requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
SQLAlchemy>=2.0.0
anthropic>=0.18.0
reportlab>=4.0.0
fpdf2>=2.0.0
psycopg2-binary>=2.9.0
